requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-mistralai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests
langchain_huggingface
streamlit